+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/).
 - Custom `includes` configuration for plugins is not correctly resolved
   if the `root` is different to `Deno.cwd()`.
 - Plugin `bundler` in Windows.
+- Plugin `postcss` with multiple includes.
 
 ## [0.25.0] - 2021-07-04
 ### Added